Fix warning

+++ pyscf/mcpdft/cmspdft.py
@@ -155,7 +155,7 @@ def e_coul_o0 (mc,ci):
     tdm1 = np.dot(trans12_tdm1_array,mo_cas.T)
     tdm1 = np.dot(mo_cas,tdm1).transpose(1,0,2)
 
-    rowscol2ind = np.zeros ((nroots, nroots), dtype=np.integer)
+    rowscol2ind = np.zeros ((nroots, nroots), dtype=int)
     rowscol2ind[(rows,col)] = list (range (pairs))
     rowscol2ind += rowscol2ind.T
     rowscol2ind[np.diag_indices(nroots)] = -1
